Iniciando git.
