+++ src/main/java/SpringBootWebApplication.java
@@ -1,0 +1,20 @@
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.support.SpringBootServletInitializer;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackages = { "es.salesianos.*" })
+public class SpringBootWebApplication extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(SpringBootWebApplication.class);
+	}
+
+	public static void main(String[] args) throws Exception {
+		SpringApplication.run(SpringBootWebApplication.class, args);
+	}
+
+}
