--- src/main/java/connections/H2Connection.java
@@ -1,61 +1,0 @@
-package connections;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class H2Connection implements ConnectionManager {
-
-	public Connection open(String jdbcUrl) {
-		Connection conn = null;
-		try {
-			Class.forName("org.h2.Driver");
-			conn = DriverManager.getConnection(jdbcUrl + ";INIT=RUNSCRIPT FROM 'classpath:scripts/create.sql'", "sa","");
-		} catch (Exception e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
-		}
-		return conn;
-	}
-
-	public Connection executeSql(Connection conn, String sql) {
-		PreparedStatement prepareStatement;
-		try {
-			prepareStatement = conn.prepareStatement(sql);
-			prepareStatement.execute(sql);
-		} catch (SQLException e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
-		}
-		return conn;
-	}
-
-	public void close(Connection conn) {
-		try {
-			conn.close();
-		} catch (SQLException e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
-		}
-	}
-
-	public void close(PreparedStatement prepareStatement) {
-		try {
-			prepareStatement.close();
-		} catch (SQLException e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
-		}
-	}
-
-	public void close(ResultSet resultSet) {
-		try {
-			resultSet.close();
-		} catch (SQLException e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
-		}
-	}
-}
