+++ src/main/java/es/salesianos/assemblers/CompanyAssembler.java
@@ -1,0 +1,16 @@
+package es.salesianos.assemblers;
+
+import java.sql.Date;
+import javax.servlet.http.HttpServletRequest;
+
+import es.salesianos.models.Company;
+
+public class CompanyAssembler {
+
+	public static Company assembleCompanyFrom(HttpServletRequest request) {
+		Company company = new Company();
+		company.setName(request.getParameter("name"));
+		company.setCreationDateFromString(request.getParameter("creationDate"));
+		return company;
+	}
+}
