+++ src/main/java/es/salesianos/models/Console.java
@@ -1,0 +1,22 @@
+package es.salesianos.models;
+
+public class Console {
+	private String name;
+	private int companyId;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public int getCompanyId() {
+		return companyId;
+	}
+
+	public void setCompanyId(int companyId) {
+		this.companyId = companyId;
+	}
+}
