+++ src/main/java/es/salesianos/repositories/ConsolesRepository.java
@@ -1,0 +1,156 @@
+package es.salesianos.repositories;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+import es.salesianos.connections.ConnectionManager;
+import es.salesianos.connections.H2Connection;
+import es.salesianos.models.Console;
+
+public class ConsolesRepository {
+	private static final String jdbcUrl = "jdbc:h2:file:./src/main/resources/test;INIT=RUNSCRIPT FROM 'classpath:scripts/create.sql'";
+	ConnectionManager manager = new H2Connection();
+
+	public Console search(Console consoleForm) {
+		Console consoleDB = null;
+		ResultSet resultSet = null;
+		PreparedStatement prepareStatement = null;
+		Connection conn = manager.open(jdbcUrl);
+		try {
+			prepareStatement = conn.prepareStatement("SELECT * FROM CONSOLE WHERE name = ?");
+			prepareStatement.setString(1, consoleForm.getName());
+			resultSet = prepareStatement.executeQuery();
+			while (resultSet.next()) {
+				consoleDB = new Console();
+				consoleDB.setName(resultSet.getString(0));
+				consoleDB.setCompanyId(resultSet.getInt(1));
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(resultSet);
+			manager.close(prepareStatement);
+			manager.close(conn);
+		}
+		return consoleDB;
+	}
+
+	public void insertConsole(Console consoleForm) {
+		Connection conn = manager.open(jdbcUrl);
+		PreparedStatement preparedStatement = null;
+		try {
+			preparedStatement = conn.prepareStatement("INSERT INTO CONSOLE (name, company)" + "VALUES (?, ?)");
+			preparedStatement.setString(1, consoleForm.getName());
+			preparedStatement.setInt(2, consoleForm.getCompanyId());
+			preparedStatement.executeUpdate();
+		} catch (SQLException e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(preparedStatement);
+			manager.close(conn);
+		}
+	}
+
+	public void update(Console consoleForm) {
+		Connection conn = manager.open(jdbcUrl);
+		manager.close(conn);
+	}
+
+	public List<Console> searchAll() {
+		List<Console> listConsole = new ArrayList<Console>();
+		Connection conn = manager.open(jdbcUrl);
+		ResultSet resultSet = null;
+		PreparedStatement prepareStatement = null;
+		try {
+			prepareStatement = conn.prepareStatement("SELECT * FROM CONSOLE");
+			resultSet = prepareStatement.executeQuery();
+			while (resultSet.next()) {
+				Console consoleInDatabase = new Console();
+				consoleInDatabase.setName(resultSet.getString(1));
+				consoleInDatabase.setCompanyId(resultSet.getInt(2));
+				listConsole.add(consoleInDatabase);
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(resultSet);
+			manager.close(prepareStatement);
+			manager.close(conn);
+		}
+		return listConsole;
+	}
+
+	public List<Console> orderByTitle() {
+		List<Console> listConsole = new ArrayList<Console>();
+		Connection conn = manager.open(jdbcUrl);
+		ResultSet resultSet = null;
+		PreparedStatement prepareStatement = null;
+		try {
+			prepareStatement = conn.prepareStatement("SELECT * FROM CONSOLE ORDER BY name ASC");
+			resultSet = prepareStatement.executeQuery();
+			while (resultSet.next()) {
+				Console consoleInDatabase = new Console();
+				consoleInDatabase.setName(resultSet.getString(1));
+				consoleInDatabase.setCompanyId(resultSet.getInt(2));
+				listConsole.add(consoleInDatabase);
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(resultSet);
+			manager.close(prepareStatement);
+			manager.close(conn);
+		}
+		return listConsole;
+	}
+
+	public void delete(Console console) {
+		Connection conn = manager.open(jdbcUrl);
+		PreparedStatement preparedStatement = null;
+		try {
+			preparedStatement = conn.prepareStatement("DELETE FROM CONSOLE WHERE name ='" + console.getName() + "'");
+			preparedStatement.setString(1, console.getName());
+			preparedStatement.executeUpdate();
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(preparedStatement);
+			manager.close(conn);
+		}
+	}
+
+	public List<Console> selectByCompany(int id) {
+		List<Console> listConsole = new ArrayList<Console>();
+		Connection conn = manager.open(jdbcUrl);
+		ResultSet resultSet = null;
+		PreparedStatement prepareStatement = null;
+		try {
+			prepareStatement = conn.prepareStatement("SELECT * FROM CONSOLE WHERE companyId = ?");
+			prepareStatement.setString(1, id + "");
+			resultSet = prepareStatement.executeQuery();
+			while (resultSet.next()) {
+				Console consoleInDatabase = new Console();
+				consoleInDatabase.setName(resultSet.getString(1));
+				consoleInDatabase.setCompanyId(resultSet.getInt(2));
+				listConsole.add(consoleInDatabase);
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		} finally {
+			manager.close(resultSet);
+			manager.close(prepareStatement);
+			manager.close(conn);
+		}
+		return listConsole;
+	}
+}
