+++ src/main/java/es/salesianos/servlets/CreateConsole.java
@@ -1,0 +1,35 @@
+package es.salesianos.servlets;
+
+import java.io.IOException;
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import es.salesianos.models.Console;
+import es.salesianos.services.ConsoleService;
+
+public class CreateConsole extends HttpServlet {
+	private ConsoleService service = new ConsoleService();
+
+	@Override
+	public void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		Console console = service.assembleUserFromRequest(req);
+		service.createNewConsoleFromRequest(console);
+		redirect(req, resp);
+	}
+
+	protected void redirect(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/CreateConsole.jsp");
+		dispatcher.forward(req, resp);
+	}
+
+	public ConsoleService getService() {
+		return service;
+	}
+
+	public void setService(ConsoleService service) {
+		this.service = service;
+	}
+}
