+++ src/main/java/es/salesianos/servlets/DeleteConsole.java
@@ -1,0 +1,41 @@
+package es.salesianos.servlets;
+
+import java.io.IOException;
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import es.salesianos.models.Console;
+import es.salesianos.services.ConsoleService;
+
+public class DeleteConsole extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+	private ConsoleService service = new ConsoleService();
+
+	@Override
+	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		Console console = service.assembleUserFromRequest(req);
+		service.deleteConsole(console);
+		loginRedirect(req, resp);
+	}
+
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.setAttribute("name", req.getParameter("name"));
+		confirmationRedirect(req, resp);
+	}
+
+	protected void confirmationRedirect(HttpServletRequest req, HttpServletResponse resp)
+			throws IOException, ServletException {
+		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("ConfirmationDeleteConsole.jsp");
+		dispatcher.forward(req, resp);
+	}
+
+	protected void loginRedirect(HttpServletRequest req, HttpServletResponse resp)
+			throws IOException, ServletException {
+		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("ListConsole.jsp");
+		dispatcher.forward(req, resp);
+	}
+}
