+++ src/main/java/es/salesianos/servlets/GamesOrderByTitle.java
@@ -1,0 +1,29 @@
+package es.salesianos.servlets;
+
+import java.io.IOException;
+import java.util.List;
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import es.salesianos.models.Game;
+import es.salesianos.services.GameService;
+
+public class GamesOrderByTitle extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+	private GameService service = new GameService();
+
+	@Override
+	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		List<Game> listAllVideoGame = service.OrderByTitle();
+		req.getSession().setAttribute("listAllVideoGame", listAllVideoGame);
+		redirect(req, resp);
+	}
+
+	protected void redirect(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/ListGame.jsp");
+		dispatcher.forward(req, resp);
+	}
+}
