+++ src/main/java/es/salesianos/servlets/SelectCompanyForGame.java
@@ -1,0 +1,29 @@
+package es.salesianos.servlets;
+
+import java.io.IOException;
+import java.util.List;
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import es.salesianos.models.Company;
+import es.salesianos.services.CompanyService;
+
+public class SelectCompanyForGame extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+	private CompanyService service = new CompanyService();
+
+	@Override
+	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		List<Company> listAllCompany = service.listAllCompany();
+		req.setAttribute("listAllCompanyVG", listAllCompany);
+		redirect(req, resp);
+	}
+
+	protected void redirect(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/ListVideoGameByCompany.jsp");
+		dispatcher.forward(req, resp);
+	}
+}
