--- src/main/java/models/Console.java
@@ -1,22 +1,0 @@
-package models;
-
-public class Console {
-	private String name;
-	private int companyId;
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public int getCompanyId() {
-		return companyId;
-	}
-
-	public void setCompanyId(int companyId) {
-		this.companyId = companyId;
-	}
-}
